refactor: move compose screen files to screens package
