+++ app/src/main/java/com/example/cupcake/ui/screens/StartScreen.kt
@@ -1,4 +1,4 @@
-package com.example.cupcake.ui
+package com.example.cupcake.ui.screens
 
 import androidx.compose.foundation.Image
 import androidx.compose.foundation.layout.Column
